Add new installation script for Debian and Ubuntu
A new bash script has been added for the installation on Debian and Ubuntu systems. This script contains basic validation and tips for usage. It distinguishes between the OS input and the source input, and log messages have been added to provide feedback. It also includes some file renaming for accuracy.