Refactor ufw installation script and fix typo in docker script
The ufw installation script has been refactored with a new function 'config_ufw' to handle common configurations. This function is called during the 'oics' strategy in the script. Moreover, a typo in the docker installation script has been fixed. The word '叫本滴' has been corrected to '脚本的', providing the correct information about the script.