Update verification function with argument flexibility
The verification function for the nginx configuration has been updated to accept an additional argument specifying the path to the docker-compose.yml file. If no argument is provided, it will default to finding the yml file in the current directory. Also, made edits in the README to reflect these changes.