Expand verify.sh with additional checks and logs
In the nginx/verify.sh script, the single source link has been replaced with two different source links for logging and checking command existence. Furthermore, the script now includes checks for whether needed files and parameters are provided and warns the user if not. It now also outputs appropriate logs in case of verification failure or success.