Enhance shellcheck and Docker validation in verify_func.sh
The verify_func.sh script in the nginx directory has been updated to include an additional shellcheck disable command in order to suppress a warning. Furthermore, a check has been added to ensure the existence of the 'docker-compose' command. If the command doesn't exist, the script flags for Docker in Docker usage. This flag is subsequently evaluated in if-statements to ascertain the optimal command to employ.