Add flexibility for Dockerfile selection in build script
The build script has been updated to accept a customizable path to the Dockerfile. A command-line argument '-f' is introduced to accept this custom path. If no custom path is provided, the script will look for a default Dockerfile in the execution directory. The README has been updated to reflect these changes.