Add build directory option to build scripts
The build scripts for gradle, maven, golang, and a new entry for writerside have been updated to add a new '-d' option to specify the directory to build from. When not specified, the current working directory is used as the build directory. The README.md file has also been updated to reflect these changes.