Add optional build directory parameter to scripts
Scripts are updated to accept an optional build directory parameter. If the provided build directory path is invalid, the scripts will exit with an appropriate error message. This allows for flexibility when choosing where the build will occur, defaulting to the Dockerfile's directory when not specified. Documentation is also updated to reflect these changes.