Update variable concatenation in docker config script
The change in this commit is about enhancing the way we concatenate variables in the docker configuration script. Specifically, we updated the method to use the more familiar and clearer syntax of enclosing the variable in curly brackets. This will help future developers easily understand the syntax and avoid misinterpretation.