🚀 Add build arg option to Docker build script
This commit introduces a new `-a` option to the `docker/build.sh` script, allowing users to pass build-time variables as arguments. The `build_arg` variable has been added to handle these parameters, and corresponding log messages have been updated to reflect this new addition. This enhances the flexibility and configurability of the Docker build process.